Task Controller and changed some logics

+++ app/Http/Controllers/Auth/UserAuthController.php
@@ -13,7 +13,7 @@ class UserAuthController extends Controller
      * Registers a user
      *
      * @param Request $request
-     * @return \Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\Response
+     * @return Response
      */
     public function register(Request $request)
     {
@@ -22,7 +22,8 @@ class UserAuthController extends Controller
             'first_name' => 'required|max:100',
             'last_name' => 'required|max:100',
             'email' => 'required|email|unique:users',
-            'password' => 'required'
+            'password' => 'required',
+            'timezone' => 'required'
         ];
 
         // validate user's data
@@ -38,7 +39,7 @@ class UserAuthController extends Controller
         $token = $user->createToken('API Token')->accessToken;
 
         // return user and its token
-        return ReturnResponse::returnJson('todo', ['user' => $user, 'token' => $token], 'success', 200);  
+        return ReturnResponse::returnJson('todo-api', ['user' => $user, 'token' => $token], true, 200);  
     }   
 
 
@@ -46,7 +47,7 @@ class UserAuthController extends Controller
      * Logs in a user
      *
      * @param Request $request
-     * @return \Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\Response
+     * @return Response
      */
     public function login(Request $request)
     {
@@ -60,13 +61,13 @@ class UserAuthController extends Controller
 
         // try to log in
         if (!auth()->attempt($data)) {
-            return ReturnResponse::returnJson('todo', ['message' => 'Incorrect Details. Please try again'], false, 404);  
+            return ReturnResponse::returnJson('todo-api', ['message' => 'Incorrect Details. Please try again'], false, 404);  
         }
 
         // introduce the access token
         $token = auth()->user()->createToken('API Token')->accessToken;
 
-        return ReturnResponse::returnJson('todo',  [
+        return ReturnResponse::returnJson('todo-api',  [
                 'user' => auth()->user(),
                 'token' => $token
             ], true, 200);  
@@ -76,21 +77,82 @@ class UserAuthController extends Controller
      * Retrieves user's data
      *
      * @param Request $request
-     * @return mixed
+     * @return Response
      */
     public function me(Request $request)
     {
         $user = $request->user();
-        return ReturnResponse::returnJson('todo',  [
+        return ReturnResponse::returnJson('todo-api',  [
                 'user' => $user
             ], true, 200); 
     }
+
+
+    /**
+     * User's data update
+     *
+     * @param Request $request
+     * @return Response
+     */
+    public function update(Request $request)    {
+
+        // introduce the validation rules
+        $validation_rules = [
+            'first_name' => 'required|max:100',
+            'last_name' => 'required|max:100',
+            'email' => 'required|email|unique:users',
+            'password' => 'required',
+            'timezone' => 'required'
+        ];
+
+        $data = [
+                'first_name' => $request->first_name,                          
+                'last_name' => $request->last_name,
+                'email' => $request->email,
+                'password' => bcrypt($request->password),
+                'timezone' => $request->timezone
+                ];
+
+        if($request->user()){ 
+            $user = User::where('id', auth()->user()->id)->update($data);     
+            return ReturnResponse::returnJson('todo-api', ["message" => 'You have successfully changed your informations'], true, 200);          
+        }else{
+            return ReturnResponse::returnJson('todo-api', ["message" => 'Changed informations is allowed only for authorized users !'], false, 401);
+        }    
+    }
+
+
+    /**
+     * User's data update timezone
+     *
+     * @param Request $request
+     * @return Response
+     */
+    public function updateTimeZone(Request $request)    {
+
+        // introduce the validation rules
+        $validation_rules = [
+            'timezone' => 'required'
+        ];
+
+        $data = [
+                'timezone' => $request->timezone
+                ];
+
+        if($request->user()){ 
+            $user = User::where('id', auth()->user()->id)->update($data);     
+            return ReturnResponse::returnJson('todo-api', ["message" => 'You have successfully changed your timezone'], true, 200);          
+        }else{
+            return ReturnResponse::returnJson('todo-api', ["message" => 'Changed informations is allowed only for authorized users !'], false, 401);
+        }    
+    }
+
 
     /**
      * Logs out user
      *
      * @param Request $request
-     * @return mixed
+     * @return Response
      */
     public function logOut(Request $request)
     {
@@ -98,11 +160,11 @@ class UserAuthController extends Controller
 
         if ($user) {
             $user->token()->revoke();
-             return ReturnResponse::returnJson('todo',  [
+             return ReturnResponse::returnJson('todo-api',  [
                 ['message' => 'You have successfully logged out']
             ], true, 200); 
         }else{
-        	return ReturnResponse::returnJson('todo',  [
+        	return ReturnResponse::returnJson('todo-api',  [
                 ['message' => 'Not authenticated.']
             ], true, 401); 
         }
@@ -112,16 +174,17 @@ class UserAuthController extends Controller
      * Delete user data
      *
      * @param Request $request
-     * @return mixed
+     * @return Response
      */
     public function delete(Request $request, $id)
     {
+        // each user can delete only himself
         $user = User::where('id', $id)->first();  
-        if ($user !== null) {      
+        if ($user !== null && $request->user()->id == $id) {      
             $user->delete();             
-            return ReturnResponse::returnJson('todo', ["message" => "delete user success"], true, 200);  
+            return ReturnResponse::returnJson('todo-api', ["message" => "You have successfully deleted all data related to you !"], true, 200);  
         }else{
-            return ReturnResponse::returnJson('todo', ["message" => "User not found"], false, 404);
+            return ReturnResponse::returnJson('todo-api', ["message" => "Data deletion is allowed only for you, pleae check your user id !"], false, 403);
         }
     }
 }
